+++ taskify-runtime-nostr/src/index.ts
@@ -2,3 +2,5 @@ export * from "./boardKeys.js";
 export * from "./relayUrls.js";
 export * from "./EventCache.js";
 export * from "./PublishCoordinator.js";
+export * from "./CursorStore.js";
+export * from "./SubscriptionManager.js";
